chore: refactor error message interpolation approach
Adds `schema` and `value` to `ValidationError`, allowing for the validators to pass extra context to the next step (message interpolation). It also removes the need for traversing the schema an extra time to calculate the correct context for each error message. 

This is why there are a lot of changes: I had to go through each validator and pass these 2 extra properties and update every test that was asserting on the errors returned by the `validateSchema` call.
## Changes
- Added `schema` and `value` fields to `ValidationError` interface to track the schema and value that caused the validation error
- Removed `getSchemaAndValueAtPath` function since schema and value are now directly available in the error object
- Simplified `addErrorMessages` function to use the schema and value directly from the error object
- Updated all validation functions to include schema and value in their error objects
- Added test utility `errorLike` to correct the validation error assertions and avoid big changes on tets.