fix(v1): use correct checkboxValue on checkbox inputs (#174)
Our previous implementation had a bug:

it only populated checkboxValue if the schema had a const and had checkbox as its input type
This is a problem because it's not taking into account the type of the schema. In other words, to mimic v0's behavior, one should set the checkboxValue to either what's defined in the const property OR to boolean, in case the type is boolean.

Added 2 groups of tests

one for boolean validation (not boolean schemas, but schemas of type boolean)
one for the fields generation